requirements.txt
langchain
langchain-community
pypdf
selenium
streamlit
beautifulsoup4
pydantic
python-dotenv
requests
groq
PyMuPDF
SpeechRecognition
gtts
pydub

